Update to latest ctrulib
This commit updates hwtests to latest ctrulib.
Some observations Citra crashes when running the GPU tests.

On ninjhax 1.1b some of the GPU tests result in failure here the
[log](https://gist.github.com/FenrisulfrX/55f1c82c1d1a89e18203).

On *hax 2.6 the Kernel tests do not run, it hardlocks the console. Still
on 2.5 a few GPU tests sometimes fail and sometimes it doesn't, however
slightly different than the failures on ninjhax 1.1b. Here's the
[log](https://gist.github.com/FenrisulfrX/19cd2303f90f4ac034b8). (GPU
tests only)